+++ package.json
@@ -2,10 +2,10 @@
   "name": "bot-hyperloot",
   "description": "",
   "version": "1.0.0",
-  "main": "server.js",
+  "main": "src/index.js",
   "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js",
+    "start": "node src/index.js",
+    "dev": "nodemon src/index.js",
     "test": "jest",
     "pretest": "npm run lint",
     "lint": "eslint '**/*.js'",
@@ -16,6 +16,7 @@
     "body-parser": "^1.18.2",
     "cors": "^2.8.4",
     "dblapi.js": "^2.2.0",
+    "debug": "^4.1.1",
     "discord.js": "^11.4.2",
     "event-stream": "3.3.4",
     "express": "^4.16.1",
@@ -25,6 +26,7 @@
     "moment": "^2.23.0",
     "mongodb": "^3.1.4",
     "mongojs": "^2.6.0",
+    "nedb": "^1.8.0",
     "node-telegram-bot-api": "0.30.0",
     "promised-csv": "^1.0.1",
     "request": "^2.88.0",
